Avoid one more warning
"comparison of integer expressions of different signedness" (int and size_t)

+++ src/aln.cpp
@@ -457,7 +457,7 @@ inline int HammingDistance(const std::string &One, const std::string &Two)
 
     int counter = 0;
 
-    for(int i=0; i<One.length(); i++) {
+    for (size_t i=0; i < One.length(); i++) {
         if (One[i] != Two[i]) counter++;
     }
